# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Core API Documentation",
        "description": "API documentation for Laravel E-Commerce Core project including Auth, Products, Categories, Cart, Orders, Coupons, Reviews/Comments, Addresses, and Admin Management.",
        "contact": {
            "name": "API Support Team",
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Lấy danh sách địa chỉ của người dùng",
                "operationId": "dc6b190d3f07db7ed032fa181dc45150",
                "responses": {
                    "200": {
                        "description": "Danh sách địa chỉ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Tạo địa chỉ mới",
                "operationId": "6344fc023028e43a5490f1c6372bd391",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "province",
                                    "district",
                                    "ward",
                                    "detail_address"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Hà Nội"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Cầu Giấy"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Dịch Vọng"
                                    },
                                    "detail_address": {
                                        "type": "string",
                                        "example": "Số 123 Xuân Thủy"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo địa chỉ thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{address}": {
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Cập nhật địa chỉ",
                "operationId": "e062d20a3459dd0b1ad80bb7ac32609e",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID địa chỉ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Hà Nội"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Cầu Giấy"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Dịch Vọng"
                                    },
                                    "detail_address": {
                                        "type": "string",
                                        "example": "Số 456 Xuân Thủy"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật địa chỉ thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Xóa địa chỉ",
                "operationId": "fc2684b35fd67400b287e6b5638d75b2",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID địa chỉ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{address}/set-default": {
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Đặt địa chỉ mặc định",
                "operationId": "b558884c88149939f02183561b9ad546",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID địa chỉ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đặt mặc định thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "Quản trị: Danh sách tất cả đơn hàng",
                "operationId": "b2cc1132ffd726ff31e8f4ed322cbeb1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số đơn hàng mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách đơn hàng toàn hệ thống"
                    },
                    "403": {
                        "description": "Không có quyền quản trị"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{order}": {
            "get": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "Quản trị: Chi tiết đơn hàng",
                "operationId": "ba919234afd9ecb389e416dac150434c",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết đơn hàng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{order}/status": {
            "put": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "Quản trị: Cập nhật trạng thái đơn hàng",
                "operationId": "8a29387fd97b9518c8315d66c54d8565",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật trạng thái thành công"
                    },
                    "422": {
                        "description": "Trạng thái không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}